Project Setup
Setup of the current project on the main branch.

Instructions follow, to change project on seperate branches.
Idea is to commit changes on "non-main" branches to move them onto the main branch upon reedit. This should solve issues, in which two versions are getting uploaded in the same time, one rewriting the other.

Until further notice, upload any changes to the main-branch