fixed the issue with the keys
fixed the issue of the keys in the upper left corner, so now they fit to the number of keys in the scene. Also redesigned the buttons, but they still dont work after the tutorial scene. There may be an issue with overlay somewhere. Also implemented a debug button to skip the cutscene in the right corner at the bottom.